% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildSubclones.R
\name{buildSubclones}
\alias{buildSubclones}
\title{Build (pure) subclones by resampling real copy-number data}
\usage{
buildSubclones(len, nbClones, bkps, regions, dataAnnotTP = NULL,
  dataAnnotN = NULL, eps = 0.25)
}
\arguments{
\item{len}{The number of loci in each subclone.}

\item{nbClones}{The number of subclones.}

\item{bkps}{A list of vectors of breakpoint positions for each subclone.}

\item{regions}{A list of altered regions for each subclone.}

\item{dataAnnotTP}{A data frame containing tumor data from which the subclone
data are sampled, with columns: \describe{ \item{\code{c}}{total copy number}
\item{\code{b}}{B allele fraction} \item{\code{region}}{type of DNA copy number
alteration} \item{\code{genotype}}{germline genotype} } If \code{NULL} (default),
this data frame is simulated using a simple model with Gaussian errors (on
the scale of log-copy-numbers).}

\item{dataAnnotN}{A data frame containing paired normal data from which the
subclone data are sampled. Same format as argument \code{dataAnnotTP}. Must not
be \code{NULL} if argument \code{dataAnnotTP} is not \code{NULL}.}

\item{eps}{A numeric value, the signal to noise ratio for simulated data.}
}
\value{
A list of locus-level data sets for each subclone. Each data set is
of the format described above for argument \code{dataAnnotTP}.
}
\description{
Build (pure) subclones by resampling real copy-number data
}
\details{
The subclones are designed to mimic \emph{intra}-tumor heterogeneity. In
particular, all of the subclones generated by a single call to this
function share the same germline genotypes. Their relative frequency
corresponds to that of the input data.

If \code{dataAnnotTP} is provided, then the output data are simply sampled
from it. Else, they are simulated as follows: (i) generate genotypes and
minor and major copy number profiles, (ii) transform them back into more
classical copy number estimates: total copy number and allele B fractions,
ie the format described above for argument \code{dataAnnotTP}.
}
\examples{
len <- 500*10  ## Number of loci
K <- 2L        ## Number of subclones
bkps <- list(c(100, 250)*10, c(150, 400)*10)
regions <- list(c("(0,1)", "(0,2)", "(1,2)"), c("(1,1)", "(0,1)", "(1,1)"))

## Resampling read copy number data
dataAnnotTP <- acnr::loadCnRegionData(dataSet="GSE13372", tumorFraction=1)
dataAnnotN <- acnr::loadCnRegionData(dataSet="GSE13372", tumorFraction=0)

datR <- buildSubclones(len=len, nbClones=K, bkps=bkps, regions=regions,
                       dataAnnotTP=dataAnnotTP, dataAnnotN=dataAnnotN)

## Simulated data
datS <- buildSubclones(len=len, nbClones=K, bkps=bkps, regions=regions)

}
