% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\docType{class}
\name{c3coClass-class}
\alias{c3coClass-class}
\title{@title Class for the object return by \code{c3co} function
 @slot BIC A numeric that is the value of BIC
 @slot PVE A numeric that is the Percentage of variation explained
 @slot res A \code{posFused} object
 @slot param List of parameters (number of subclones, penalty coefficients)
 @slot bkp A list of breakpoints for each chromosome
 @exportClass c3coClass}
\description{
@title Class for the object return by \code{c3co} function
 @slot BIC A numeric that is the value of BIC
 @slot PVE A numeric that is the Percentage of variation explained
 @slot res A \code{posFused} object
 @slot param List of parameters (number of subclones, penalty coefficients)
 @slot bkp A list of breakpoints for each chromosome
 @exportClass c3coClass
}

