% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{c3coFit-class}
\alias{c3coFit-class}
\title{@title Class for the object create by \code{c3coFit} function}
\description{
@slot bkp A list of breakpoints for each chromosome.
}
\details{
@slot segDat A list that contains segmented data.

@slot fit A List of [\code{\linkS4class{posFused}}] objects.

@exportClass c3coClass
}
