% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\docType{methods}
\name{createZdf}
\alias{createZdf}
\alias{createZdf,c3coClass-method}
\title{The method createZdf}
\usage{
createZdf(this, minMaxPos, chromosomes, var = "TCN")

\S4method{createZdf}{c3coClass}(this, minMaxPos, chromosomes, var = "TCN")
}
\arguments{
\item{this}{an object from the following class: [\code{\linkS4class{c3coClass}}]}

\item{minMaxPos}{Matrix that contains min and max position for each chromosome}

\item{chromosomes}{A vector that contains the focused chromosomes}

\item{var}{TCN, Minor or Major}
}
\value{
A data frame to plot Latent profiles with ggplot
}
\description{
Create a data frame to plot Subclones
\itemize{
\item{createZdf(this)}{this is an object from class  [\code{\linkS4class{c3coClass}}].}
}
}

