% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Facetswrapper.R
\name{Facetswrapper}
\alias{Facetswrapper}
\title{Transforms Facets data and performs a joint segmentation}
\usage{
Facetswrapper(pathFacets, stat)
}
\arguments{
\item{pathFacets}{The path to load Facets data.}

\item{stat}{"TCN or "C1C2" paramater to segment the data.
If \code{stat == TCN}, the segmentation will be done on TCN only.}
}
\value{
A list which contains the breakpoints by chromosome and also the
binning of TCN, C1, and C2.
}
\description{
Transforms Facets data and performs a joint segmentation
}
\examples{
if (require("facets", quietly=TRUE)) {
pathFacets <- system.file("extdata", package="facets")
segDat <- Facetswrapper(pathFacets, stat="TCN")
\dontrun{
resc3co <- c3co(NULL, segDat=segDat)
}
}

}
