% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.W.R
\name{get.W}
\alias{get.W}
\title{Get the matrix of weights W for fixed values of the subclones Z and data Y}
\usage{
get.W(Zt, Y)
}
\arguments{
\item{Zt}{The transpose of Z where Z is a matrix with K rows (number of subclones) and J columns (number of segments)}

\item{Y}{a matrix with n rows (number of samples) and J columns (number of segments)}
}
\value{
a matrix with n rows (number of samples) and K columns (number of subclones)
}
\description{
Optimisation is done by solving a least square problem with inequality constraint thanks to \pkg{lsei} package
}
\examples{
nbSegments <- 11L
nbClones   <- 4L

Z <- matrix(1, nrow = nbClones, ncol = nbSegments)
Z[2, 2] <- 2
Z[3, 5:6] <- 2
Z[4, 9:10] <- 2

W <- diag(rep(1, times = nrow(Z)))
E <- matrix(rnorm(nrow(W)*ncol(Z), sd = 0), nrow = nrow(W), ncol = ncol(Z))
Y <- W \%*\% Z + E
W <- c3co:::get.W(t(Z), Y)

}
