% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.Z.R
\name{get.Z}
\alias{get.Z}
\title{Get the matrix of subclones Z for fixed values of the weights W}
\usage{
get.Z(Y, lambda, W, WtWm1)
}
\arguments{
\item{Y}{a matrix with n rows (number of samples) and J columns (number of segments)}

\item{lambda}{a positive scalar tuning the penalty level for fusion}

\item{W}{a matrix with n rows (number of samples) and K columns (number of subclones)}

\item{WtWm1}{a K x K square matrix (with K the number of subclones), precomputed to save time}
}
\value{
The transpose of Z, where Z a matrix with K rows (number of subclones) and
J columns (number of segments)
}
\description{
Optimisation is done by solving a L1-penalized least-square problem thanks to the \pkg{glmnet} package
}
\examples{
nbSegments <- 11
nbClones   <- 4

Z <- matrix(1, nrow = nbClones, ncol = nbSegments)
Z[2, 2] <- 2
Z[3, 5:6] <- 2
Z[4, 9:10] <- 2

W <- diag(rep(1, times = nrow(Z)))
E <- matrix(rnorm(nrow(W)*ncol(Z), sd = 0), nrow = nrow(W), ncol = ncol(Z))
WtWm1 <- diag(rep(1, times = nrow(Z)))
Y <- W \%*\% Z + E

Zt <- c3co:::get.Z(Y, lambda = 0.01, W, WtWm1)

}
