% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.Zt.R
\name{get.Zt}
\alias{get.Zt}
\title{Get the matrix of subclones Z for fixed values of the weights W}
\usage{
get.Zt(Y, lambda, W, WtWm1)
}
\arguments{
\item{Y}{a matrix with n rows (number of samples) and J columns (number of segments)}

\item{lambda}{a positive scalar tuning the penalty level for fusion}

\item{W}{a matrix with n rows (number of samples) and K columns (number of subclones)}

\item{WtWm1}{a K x K square matrix (with K the number of subclones), precomputed to save time}
}
\value{
The transpose of Z, where Z a matrix with K rows (number of subclones) and
J columns (number of segments)
}
\description{
Optimisation is done by solving a L1-penalized least-square problem thanks to the \pkg{glmnet} package
}
\examples{
J <- 11L  ## Number of segments
K <- 4L   ## Number of subclones
n <- 4L   ## Number of samples

Z <- matrix(1, nrow = K, ncol = J)
Z[2,    2] <- 2
Z[3,  5:6] <- 2
Z[4, 9:10] <- 2

W <- matrix(0, nrow = n, ncol = K)
W[1,1] <- W[2,2] <- W[3,3] <- W[4,4] <- 1
WtWm1 <- tcrossprod(backsolve(qr.R(qr(W)), x=diag(K)))

E <- matrix(rnorm(n*J, sd = 0.1), nrow = n, ncol = J)

Y <- W \%*\% Z + E

Zthat <- c3co:::get.Zt(Y, lambda = 0.01, W = W, WtWm1 = WtWm1)

}
