% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWeightMatrix.R
\name{getWeightMatrix}
\alias{getWeightMatrix}
\title{Generate weight matrix}
\usage{
getWeightMatrix(prop.max, prob.min, nb.arch, nb.samp, sparse.coeff = 0.5,
  contam.coeff = 0.6, contam.max = 30)
}
\arguments{
\item{prop.max}{An integer which is the maximim proportion for a present archetype}

\item{prob.min}{An integer which is the minimum proportion  for a present archetype}

\item{nb.arch}{An interger which is the number of archetypes (number of columns)}

\item{nb.samp}{An interger which is the number of samples (number of rows)}

\item{sparse.coeff}{A numeric between 0 and 1 which control the sparsity by rows}
}
\value{
A matrix of weights (the sum of rows is equal to 1)
}
\description{
Generate weight matrix
}
\examples{
M <- getWeightMatrix(70,20, 7, 30)
}

