% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InCaSCN.R
\name{InCaSCN}
\alias{InCaSCN}
\title{Cancer subclone Inference function}
\usage{
InCaSCN(dat, lambda1.grid = NULL, lambda2.grid = NULL,
  nb.arch.grid = 2:(length(dat) - 1), stat = "C1C2")
}
\arguments{
\item{dat}{A list of data frame for each patient containing the total copy number \code{tcn}, the mirrored B allele fraction \code{dh}.}

\item{lambda1.grid}{A grid of real numbers which is the penalty coefficients for the fused lasso on the minor copy number dimension}

\item{lambda2.grid}{A grid of real numbers which is the penalty coefficients for the fused lasso on the major copy number dimension}

\item{nb.arch}{A vector of integers which is the number of archetypes in the model}
}
\value{
A list of archetypes (\code{Z} the total copy number matrix,\code{Z1} the minor copy number matrix and \code{Z2} the major copy number matrix), matrix weight \code{W} and the reconstructed minor and major copy numbers.
}
\description{
Cancer subclone Inference function
}
\examples{
dataAnnotTP <- loadCnRegionData(dataSet="GSE11976", tumorFrac=1)
dataAnnotN <- loadCnRegionData(dataSet="GSE11976", tumorFrac=0)
len <- 500*10
nbClones <- 3
bkps <- list(c(100,250)*10, c(150,400)*10,c(150,400)*10)
regions <-list(c("(0,3)", "(0,2)","(1,2)"), c("(1,1)", "(0,1)","(1,1)"), c("(0,2)", "(0,1)","(1,1)"))
datSubClone <- buildSubclones(len, dataAnnotTP, dataAnnotN, nbClones, bkps, regions)
M <- getWeightMatrix(100,0, 3, 15, sparse.coeff=0.7, contam.coeff=0.6, contam.max=2)
dat <- apply(M, 1, mixSubclones, subClones=datSubClone, fracN=NULL)
casResC1C2 <- InCaSCN(dat)
casRes <- InCaSCN(dat, stat="TCN")
}

