% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSCBSwrapper.R
\name{loadPSCBSdata}
\alias{loadPSCBSdata}
\title{Function to load PSCBS data and transform them to c3co format}
\usage{
loadPSCBSdata(pathSegPSCBS, pattern = NULL)
}
\arguments{
\item{pathSegPSCBS}{The path to load PSCBS data.}

\item{pattern}{if you have several patients in the pathSegPSCBS directory.}
}
\value{
A data frame under PSCBS format
}
\description{
Function to load PSCBS data and transform them to c3co format
}

