% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixSubclone.R
\name{mixSubclones}
\alias{mixSubclones}
\title{Create a mixture of subclones}
\usage{
mixSubclones(subClones, weights, fracN = NULL)
}
\arguments{
\item{subClones}{The subclones used to create the mixture ideally from the function \code{buildSubClones}.}

\item{weights}{A vector of weights in percentage}
}
\value{
The mixture
}
\description{
Create a mixture of subclones
}
\examples{
dataAnnotTP <- loadCnRegionData(dataSet="GSE11976", tumorFrac=1)
dataAnnotN <- loadCnRegionData(dataSet="GSE11976", tumorFrac=0)
len <- 500*10
nbClones <- 2
bkps <- list(c(100,250)*10, c(150,400)*10)
regions <-list(c("(0,3)", "(0,2)","(1,2)"), c("(1,1)", "(0,1)","(1,1)"))
datSubClone <- buildSubclones(len, dataAnnotTP, dataAnnotN, nbClones, bkps, regions)
mixture <- mixSubclones(datSubClone, c(20,30))
}

