% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/modelFitStats.R
\docType{methods}
\name{modelFitStats}
\alias{modelFitStats}
\alias{modelFitStats,posFused-method}
\title{Compute statistics of a c3co model}
\usage{
modelFitStats(object)

\S4method{modelFitStats}{posFused}(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{posFused}}}
}
\description{
Compute statistics of a c3co model
}
