% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentData.R
\name{nbrOfChromosomes}
\alias{nbrOfChromosomes}
\title{Get number of chromosomes}
\usage{
nbrOfChromosomes(x, ...)
}
\arguments{
\item{x}{A segmentation.}

\item{\ldots}{Optional arguments.}
}
\description{
Get number of chromosomes
}
