% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentData.R
\name{nbrOfSamples}
\alias{nbrOfSamples}
\title{Get number of samples}
\usage{
nbrOfSamples(x, ...)
}
\arguments{
\item{x}{A segmentation.}

\item{\ldots}{Optional arguments.}
}
\description{
Get number of samples
}
