% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentData.R
\name{nbrOfSegments}
\alias{nbrOfSegments}
\title{Get number of segments}
\usage{
nbrOfSegments(x, ...)
}
\arguments{
\item{x}{A segmentation.}

\item{\ldots}{Optional arguments.}
}
\description{
Get number of segments
}
