% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\docType{class}
\name{posFused-class}
\alias{posFused-class}
\title{@title Class for the object return by \code{positiveFusedLasso} function
 @slot S A list containing Total copy number, minor copy number and major copy number inferred by \code{positiveFusedLasso}
 @slot W A matrix containing weights inferred by \code{positiveFusedLasso}
 @slot E A list containing estimates of minor copy number and major copy number inferred by \code{positiveFusedLasso}
 @exportClass posFused}
\description{
@title Class for the object return by \code{positiveFusedLasso} function
 @slot S A list containing Total copy number, minor copy number and major copy number inferred by \code{positiveFusedLasso}
 @slot W A matrix containing weights inferred by \code{positiveFusedLasso}
 @slot E A list containing estimates of minor copy number and major copy number inferred by \code{positiveFusedLasso}
 @exportClass posFused
}

