% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positiveFusedLasso.R
\name{positiveFusedLasso}
\alias{positiveFusedLasso}
\title{Positive fused lasso function}
\usage{
positiveFusedLasso(Y1, Y2, Z1, Z2, lambda1, lambda2, eps = 0.1,
  max.iter = 50, warn = FALSE, verbose = FALSE)
}
\arguments{
\item{Y1}{A matrix containing the segmented minor copy number
(\code{n} patients in row and \code{L} segments in columns)}

\item{Y2}{A matrix containing the segmented major copy number
(\code{n} patients in row and \code{L} segments in columns)}

\item{Z1}{A \code{L} x \code{p} matrix containing the \code{L}
minor copy numbers of the \code{p} initial latent feature estimates}

\item{Z2}{A \code{L} x \code{p} matrix containing the \code{L}
major copy numbers of the \code{p} initial latent feature estimates}

\item{lambda1}{A real number, the coefficient for the fused penalty
for minor copy numbers}

\item{lambda2}{A real number, the coefficient for the fused penalty
for major copy numbers}

\item{eps}{criterion to stop algorithm (when W do not change
sqrt(sum((W-W.old)^2) < eps)}

\item{max.iter}{maximum number of iterations of the algorithm}

\item{warn}{issue a warning if Z1 <= Z2 is not always satisfied?
Defaults to FALSE}

\item{verbose}{if you want to print some information during running}
}
\value{
An object of class [\code{\linkS4class{posFused}}]
}
\description{
Positive fused lasso function
}
\examples{
dataAnnotTP <- acnr::loadCnRegionData(dataSet="GSE11976", tumorFrac=1)
dataAnnotN <- acnr::loadCnRegionData(dataSet="GSE11976", tumorFrac=0)
len <- 500*10
nbClones <- 3
set.seed(2)
bkps <- list(c(100, 250)*10, c(150, 400)*10, c(150, 400)*10)
regions <- list(c("(0,3)", "(0,1)", "(1,2)"),
c("(1,1)", "(0,1)", "(1,1)"), c("(0,2)", "(0,1)", "(1,1)"))
datSubClone <- buildSubclones(len, nbClones, bkps, regions, dataAnnotTP, dataAnnotN)
M <- rSparseWeightMatrix(15, 3, 0.4)
simu <- mixSubclones(subClones=datSubClone, M)
seg <- segmentData(simu)
lambda <- 0.01
Z <- initializeZ(t(seg$Y1), t(seg$Y2), p=3)
res <- positiveFusedLasso(t(seg$Y1), t(seg$Y2), Z$Z1, Z$Z2,
                          lambda1=lambda, lambda2=lambda, verbose=TRUE)
res
resC <- positiveFusedLasso(t(seg$Y1+seg$Y2), Y2=NULL, Z1=Z$Z, Z2=NULL,
                           lambda1=lambda, lambda2=lambda)
resC

}
