% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSparseWeightMatrix.R
\name{rSpMatrix}
\alias{rSpMatrix}
\title{Generate sparse matrix with at least one non-zero element by rows and by columns}
\usage{
rSpMatrix(nrow, ncol, sparse.coeff)
}
\arguments{
\item{nrow}{An integer that is the number of rows.}

\item{ncol}{An integer that is the number of columns.}

\item{sparse.coeff}{A numeric in \eqn{[0,1]} that control the sparsity by
rows (the proportion of non-zero entries among all matrix entries).
Non-zero entries is define by \code{nnz = sparse.coeff * nrow * ncol} and
need to be larger than \code{nrow} and \code{ncol}.}
}
\value{
A matrix of weights (the sum of rows is equal to 1).
}
\description{
Generate sparse matrix with at least one non-zero element by rows and by columns
}
\examples{
M <- rSpMatrix(nrow=10L, ncol=7L, sparse.coeff=0.5)

}
