% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\docType{methods}
\name{showC3coClass}
\alias{showC3coClass}
\alias{showC3coClass,c3coClass-method}
\title{The method showC3co}
\usage{
showC3coClass(this)

\S4method{showC3coClass}{c3coClass}(this)
}
\arguments{
\item{this}{an object from the following class: [\code{\linkS4class{c3coClass}}]}
}
\value{
nothing
}
\description{
Print the message that is contained in slots BIC, PVE, res, param and bkp.
\itemize{
\item{showC3coClass(this)}{this is an object from class  [\code{\linkS4class{c3coClass}}].}
}
}

