% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{showC3coFit}
\alias{showC3coFit}
\alias{showC3coFit,c3coFit-method}
\title{Display the content of an object of class c3coFit}
\usage{
showC3coFit(this)

\S4method{showC3coFit}{c3coFit}(this)
}
\arguments{
\item{this}{An object of class \code{\linkS4class{c3coFit}}}
}
\value{
nothing
}
\description{
Display the content of an object of class c3coFit
}

