% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{showPosFused}
\alias{showPosFused}
\alias{showPosFused,posFused-method}
\title{Display the content of an object of class posFused}
\usage{
showPosFused(this)

\S4method{showPosFused}{posFused}(this)
}
\description{
Display the content of an object of class posFused
}
\section{Slots}{

\describe{
\item{\code{this}}{an object from the following class: [\code{\linkS4class{posFused}}]}
}}

