% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildSubclones.R
\name{buildSubclones}
\alias{buildSubclones}
\title{Build several subclones}
\usage{
buildSubclones(len, nbClones, bkps, regions, dataAnnotTP = NULL,
  dataAnnotN = NULL, eps = 0.25)
}
\arguments{
\item{len}{The number of loci in each subclone}

\item{nbClones}{The number of subclones}

\item{bkps}{A list of breakpoints for each subclone}

\item{regions}{A list of altered regions for each subclone}

\item{dataAnnotTP}{A \code{data.frame} containing tumor data from which the 
subclone data are sampled, with columns: \describe{ \item{c}{total copy 
number} \item{b}{B allele fraction} \item{region}{type of DNA copy number 
alteration} \item{genotype}{germline genotype}}. If \code{NULL} (the 
default), this \code{data.frame} is simulated using a simple model with 
Gaussian errors (on the scale of log-copy-numbers).}

\item{dataAnnotN}{A data frame containing paired normal data from which the 
subclone data are sampled. Same format as argument \code{dataAnnotTP}. Must
not be \code{NULL} if argument \code{dataAnnotTP} is not \code{NULL}}

\item{eps}{A numeric value, the signal to noise ratio for simulated data.}
}
\value{
A list of locus-level data for each subclone
}
\description{
Build several subclones
}
\details{
The subclones are designed to mimic intra-tumor heterogeneity. In 
  particular, all of the subclones generated by a single call to this
  function share the same germline genotypes. Their relative frequency
  corresponds to that of the input data.
  
  We start by generating genotypes and minor and major copy number profiles, 
  and then transform them back into more classical copy number estimates: 
  total copy number and allele B fractions
}
\examples{
len <- 500*10
nbClones <- 2
bkps <- list(c(100, 250)*10, c(150, 400)*10)
regions <- list(c("(0,3)", "(0,2)", "(1,2)"), c("(1,1)", "(0,1)", "(1,1)"))

## Resampling read copy number data
dataAnnotTP <- acnr::loadCnRegionData(dataSet="GSE11976", tumorFraction=1)
dataAnnotN <- acnr::loadCnRegionData(dataSet="GSE11976", tumorFraction=0)

datR <- buildSubclones(len, nbClones, bkps, regions,
                              dataAnnotTP, dataAnnotN)

## Simulated data
datS <- buildSubclones(len, nbClones, bkps, regions)

}
