% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buidlSubclones.R
\name{buildSubclones}
\alias{buildSubclones}
\title{Build several subclones}
\usage{
buildSubclones(len, dataAnnotTP, dataAnnotN, nbClones, bkps = list(),
  regions = list())
}
\arguments{
\item{len}{The number of probes in subclones.}

\item{dataAnnotTP}{A data frame containing the data used for the resampling. Need the total copy number in tumor \code{c}, \code{b} the B allele fraction in tumor, the type of alteration in \code{region} and the genotype in reference \code{genotype}}

\item{dataAnnotN}{A data frame containing the data used for the resampling. Need the total copy number in normal \code{c}, \code{b} the B allele fraction in normal, the type of alteration in \code{region} and the genotype in reference \code{genotype}}

\item{nbClones}{The number of subclones}

\item{bkps}{The list of breakpoints by subclones}

\item{regions}{The list of altered regions by subclones}
}
\value{
The list of subclone.
}
\description{
Build several subclones
}
\examples{
dataAnnotTP <- acnr::loadCnRegionData(dataSet="GSE11976", tumorFrac=1)
dataAnnotN <- acnr::loadCnRegionData(dataSet="GSE11976", tumorFrac=0)
len <- 500*10
nbClones <- 2
bkps <- list(c(100,250)*10, c(150,400)*10)
regions <-list(c("(0,3)", "(0,2)","(1,2)"), c("(1,1)", "(0,1)","(1,1)"))
datSubClone <- buildSubclones(len, dataAnnotTP, dataAnnotN, nbClones, bkps, regions)
dataAnnotTP <- acnr::loadCnRegionData(dataSet="GSE13372", tumorFrac=1)
dataAnnotN <- acnr::loadCnRegionData(dataSet="GSE13372", tumorFrac=0)
len <- 500*10
nbClones <- 2
bkps <- list(c(100,250)*10, c(150,400)*10)
regions <-list(c("(0,1)", "(0,2)","(1,2)"), c("(1,1)", "(0,1)","(1,1)"))
datSubClone <- buildSubclones(len, dataAnnotTP, dataAnnotN,  nbClones, bkps, regions)
}

