% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c3co.R
\name{c3co}
\alias{c3co}
\title{Cancer subclone inference}
\usage{
c3co(dat, parameters.grid = NULL, stat = c("C1C2", "TCN"), pathSeg = NULL,
  segDat = NULL, ..., verbose = FALSE)
}
\arguments{
\item{dat}{A list of data frames for each patient, each of them of the form
\describe{
  \item{tcn}{Total copy number}
  \item{dh}{Mirrored B allele fraction}
  \item{pos}{Position on the genome}
  \item{chr}{Chromosome}
}}

\item{parameters.grid}{A list composed of two vectors named \code{lambda1}
and \code{lambda2} of real numbers which are the penalty coefficients for
the fused lasso on the minor and major copy number dimension and a vector
named \code{nb.arch} of integers which is the number of archetypes in the
model}

\item{stat}{TCN or C1C2}

\item{segDat}{Either a path to the file that contains segmentation (.rds file), 
by default \code{NULL} or a file that contains segmentation 
(for example use segmentData function in the package)}

\item{\dots}{Further arguments to be passed to \code{\link{fitC3co}}}

\item{verbose}{A logical value indicating whether to print extra
information. Defaults to FALSE}
}
\value{
An object of class [\code{\linkS4class{c3coFit}}]
}
\description{
Cancer subclone inference
}
\examples{
dataAnnotTP <- acnr::loadCnRegionData(dataSet="GSE11976", tumorFrac=1)
dataAnnotN <- acnr::loadCnRegionData(dataSet="GSE11976", tumorFrac=0)
len <- 500*10
nbClones <- 3
bkps <- list(c(100, 250)*10, c(150, 400)*10, c(150, 400)*10)
regions <- list(c("(0,3)", "(0,2)", "(1,2)"),
c("(1,1)", "(0,1)", "(1,1)"), c("(0,2)", "(0,1)", "(1,1)"))
datSubClone <- buildSubclones(len, dataAnnotTP, dataAnnotN,
                              nbClones, bkps, regions)
M <- rSparseWeightMatrix(15, 3, 0.5)
dat <- mixSubclones(subClones=datSubClone, M)
l1 <- seq(from=1e-6, to=1e-4, length.out=10)
parameters.grid <- list(lambda=l1, nb.arch=2:6)
res <- c3co(dat, parameters.grid)
resC <- c3co(dat, stat="TCN", parameters.grid)
}
