% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{c3coFit-class}
\alias{c3coFit-class}
\title{Class for the object create by \code{\link[=fitC3co]{fitC3co()}}}
\description{
Class for the object create by \code{\link[=fitC3co]{fitC3co()}}
}
\section{Slots}{

\describe{
\item{\code{bkp}}{A list of breakpoints for each chromosome.}

\item{\code{segDat}}{A list that contains segmented data.}

\item{\code{config}}{A list containing PVE, BIC and parameters for all tested
models and for the best model.}

\item{\code{fit}}{A list of \link[=posFused-class]{posFused} objects.}
}}

