% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{c3coFit-class}
\alias{c3coFit-class}
\title{Class for the object create by \code{c3coFit} function}
\description{
Class for the object create by \code{c3coFit} function
}
\section{Slots}{

\describe{
\item{\code{bkp}}{A list of breakpoints for each chromosome.}

\item{\code{segDat}}{A list that contains segmented data.}

\item{\code{config}}{A data.frame containing PVE, BIC and parameters of models}

\item{\code{fit}}{A List of [\code{\linkS4class{posFused}}] objects.}
}}

