% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{createZdf}
\alias{createZdf}
\alias{createZdf,c3coFit-method}
\title{The method createZdf}
\usage{
createZdf(this, minMaxPos, chromosomes, var = "TCN", idxBest)

\S4method{createZdf}{c3coFit}(this, minMaxPos, chromosomes, var = "TCN",
  idxBest)
}
\arguments{
\item{this}{an object from the following class: [\code{\linkS4class{c3coFit}}]}

\item{minMaxPos}{Matrix that contains min and max position for each chromosome}

\item{chromosomes}{A vector that contains the focused chromosomes}

\item{var}{TCN, Minor or Major}

\item{idxBest}{a integer that is the best fitting of the data}
}
\value{
A data frame to plot Latent profiles with ggplot
}
\description{
Create a data frame to plot Subclones
\itemize{
\item{createZdf(this)}{this is an object from class  [\code{\linkS4class{c3coFit}}].}
}
}

