% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{createZdf}
\alias{createZdf}
\alias{createZdf,c3coFit-method}
\title{Create a data frame to plot subclones}
\usage{
createZdf(this, chromosomes, idxBest, var = "TCN", verbose = FALSE)

\S4method{createZdf}{c3coFit}(this, chromosomes, idxBest, var = c("TCN",
  "Minor", "Major"), verbose = FALSE)
}
\arguments{
\item{this}{An object of class \link[=c3coFit-class]{c3coFit}.}

\item{chromosomes}{A vector that contains the focused chromosomes.}

\item{idxBest}{A integer that is the best fitting of the data.}

\item{var}{\code{"TCN"}, \code{"Minor"} or \code{"Major"}.}

\item{verbose}{A logical value indicating whether to print extra information.
Defaults to \code{FALSE}.}
}
\value{
A data frame to plot Latent profiles with \pkg{ggplot2}.
}
\description{
Create a data frame to plot subclones
}
