% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{createZdf}
\alias{createZdf}
\title{Function to plot PVE}
\usage{
createZdf(minMaxPos, dataBest, chromosomes, var = "TCN")
}
\arguments{
\item{minMaxPos}{Matrix that contains min and max position for each chromosome}

\item{dataBest}{A list from output of InCaSCN}

\item{var}{TCN, Minor or Major}

\item{chromsomes}{A vector that contains the focused chromosomes}
}
\value{
A data frame to plot Latent profiles with ggplot
}
\description{
Function to plot PVE
}

