% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{createZdf}
\alias{createZdf}
\alias{createZdf,c3coFit-method}
\title{Create a data frame to plot Subclones}
\usage{
createZdf(this, chromosomes, idxBest, var = "TCN", verbose = FALSE)

\S4method{createZdf}{c3coFit}(this, chromosomes, idxBest, var = c("TCN",
  "Minor", "Major"), verbose = FALSE)
}
\arguments{
\item{this}{An object of class \code{\linkS4class{c3coFit}}}

\item{chromosomes}{A vector that contains the focused chromosomes}

\item{idxBest}{a integer that is the best fitting of the data}

\item{var}{TCN, Minor or Major}

\item{verbose}{print information}
}
\value{
A data frame to plot Latent profiles with ggplot
}
\description{
Create a data frame to plot Subclones
}

