% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Facetswrapper.R
\name{Facetswrapper}
\alias{Facetswrapper}
\title{Transforms FACETS data and performs a joint segmentation}
\usage{
Facetswrapper(path, stat)
}
\arguments{
\item{path}{The path to a directory containing FACETS data files
(\file{*.csv.gz}) produced by the \pkg{facets} package.}

\item{stat}{\code{"TCN"} or \code{"C1C2"} parameter to segment the data.
If \code{stat == "TCN"}, the segmentation will be done on TCN only.}
}
\value{
A list which contains the breakpoints by chromosome and also the
binning of TCN, C1, and C2.
}
\description{
Transforms FACETS data and performs a joint segmentation
}
\examples{
if (require("facets", quietly=TRUE)) {
\dontrun{
path <- system.file("extdata", package = "facets")
segDat <- Facetswrapper(path, stat = "TCN")
print(segDat)
resc3co <- c3co(NULL, segDat = segDat)
}
}

}
