% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.W.R
\name{get.W}
\alias{get.W}
\title{Get the matrix of weights W for fixed values of the subclones Z and data Y}
\usage{
get.W(Zt, Y, nu = 1e-08)
}
\arguments{
\item{Zt}{The transpose of Z where Z is a matrix with K rows (number of subclones) and M x J columns (number of signal x number of segments)}

\item{Y}{a matrix with n rows (number of samples) and M x J columns (number of signal x number of segments)}

\item{nu}{positive double standing for ridge regularization in the quadratic programming solver}
}
\value{
a matrix with n rows (number of samples) and K columns (number of subclones)
}
\description{
Optimisation is done by solving a least square problem with equality + inequality constraint thanks to the \pkg{quadprog} package
}
\examples{
## Example with M = 1 signal
J <- 11L  ## Number of segments
K <- 4L   ## Number of subclones
n <- 4L   ## Number of samples

Z <- matrix(1, nrow = K, ncol = J)
Z[2,    2] <- 2
Z[3,  5:6] <- 2
Z[4, 9:10] <- 2

W <- matrix(0, nrow = n, ncol = K)
W[1,1] <- W[2,2] <- W[3,3] <- W[4,4] <- 1

E <- matrix(rnorm(n*J, sd = 0.1), nrow = n, ncol = J)

Y <- W \%*\% Z + E

What <- c3co:::get.W(t(Z), Y)

}
