% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeZ.R
\name{initializeZ}
\alias{initializeZ}
\title{Initialization of the latent features of the c3co model}
\usage{
initializeZ(Y1, Y2 = NULL, p = min(dim(Y1)), flavor = c("hclust", "nmf",
  "archetypes", "svd", "subsampling"), stat = c("C1+C2", "C1", "C2"),
  verbose = FALSE)
}
\arguments{
\item{Y1}{A matrix containing the segmented minor copy number (\code{n}
patients in row and \code{L} segments in columns)}

\item{Y2}{A matrix containing the segmented major copy number (\code{n}
patients in row and \code{L} segments in columns)}

\item{p}{An integer value, the number of latent features in the model.
Defaults to \code{min(dim(Y1))}}

\item{flavor}{A character value specifying how initialization is perfomed.
Defaults to \code{"hclust"}. See Details}

\item{stat}{Statistic used to perform initialization. Should be either
"C1+C2", "C1", or "C2"}

\item{verbose}{A logical value indicating whether to print extra information.
Defaults to FALSE}
}
\value{
A list with two components: \describe{ \item{Z1}{A \code{L} x
  \code{p} matrix, the initial value for the \code{L} minor copy numbers of
  the \code{p} latent features} \item{Z2}{A \code{L} x \code{p} matrix, the
  initial value for the \code{L} major copy numbers of the \code{p} latent
  features} }
}
\description{
Initialization of the latent features of the c3co model
}
\details{
The latent features (LF) are inferred as follows according to the
  value of argument 'flavor':

  If \code{flavor=="hclust"} (the default), the LF are centers of clusters
  derived by hierarchical agglomerative clustering on the Euclidean distance
  between the input copy number profiles, and using Ward linkage
  (\code{\link[stats]{hclust}}).

  If \code{flavor=="nmf"}, the LF are the _coefficients_ of the non-negative
  matrix factorization in \code{p} of the input copy number profiles.

  If \code{flavor=="svd"}, the LF are the first \code{p} right singular
  vectors of the singular value decomposition of the input copy number
  profiles. The flavor is not recommended as it may produce matrices
  with non-positive entries

  If  \code{flavor=="archetypes"}, the LF are defined using archetypal
  analysis.

  If  \code{flavor=="subsampling"}, the LF are chosen at random among existing
  profiles.
}
\examples{
dataAnnotTP <- acnr::loadCnRegionData(dataSet="GSE11976", tumorFrac=1)
dataAnnotN <- acnr::loadCnRegionData(dataSet="GSE11976", tumorFrac=0)
len <- 500*10
nbClones <- 3
bkps <- list(c(100, 250)*10, c(150, 400)*10, c(150, 400)*10)
regions <- list(c("(0,3)", "(0,1)", "(1,2)"),
                c("(1,1)", "(0,1)", "(1,1)"),
                c("(0,2)", "(0,1)", "(1,1)"))
datSubClone <- buildSubclones(len, nbClones, bkps, regions, dataAnnotTP, dataAnnotN)
M <- rSparseWeightMatrix(12, nbClones, 0.90)
simu <- mixSubclones(subClones=datSubClone, M)
seg <- segmentData(simu)
Y1 <- t(seg$Y1)
Y2 <- t(seg$Y2)

resH <- initializeZ(Y1, Y2, p=nbClones)  ## corresponds to flavor "hclust")
resNMF <- initializeZ(Y1, Y2, p=nbClones, flavor="nmf")
\dontrun{
## often fails because of singularities:
resArch <- initializeZ(Y1, Y2, p=nbClones, flavor="archetypes")
}
resSVD <- initializeZ(Y1, Y2, p=nbClones, flavor="svd")
resC <- initializeZ(Y1, Y2, p=nbClones, flavor="subsampling")

resNMF1 <- initializeZ(Y1, p=nbClones, flavor="nmf")

}
\references{
Gaujoux R and Seoighe C (2010). A flexible R package for
  nonnegative matrix factorization. BMC Bioinformatics, 11(1), pp. 367.

Cutler A and Breiman L. (1994) Archetypal analysis.
  Technometrics, 36(4):338-3474.
}
