% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeZ.R
\name{initializeZ}
\alias{initializeZ}
\title{Initialization of the c3co model parameters}
\usage{
initializeZ(Y1, Y2 = NULL, nb.arch = ncol(Y1), init.random = FALSE,
  flavor = c("C1+C2", "C1", "C2"), verbose = FALSE)
}
\arguments{
\item{Y1}{A matrix containing the segmented minor copy number
(\code{n} patients in row and \code{L} segments in columns)}

\item{Y2}{A matrix containing the segmented major copy number
(\code{n} patients in row and \code{L} segments in columns)}

\item{nb.arch}{An integer which is the number of archetypes in the model}

\item{init.random}{if you want to use random initialization set parameter
to TRUE}

\item{flavor}{Statistic used to perform 'hclust' initialization.
Should be either "C1+C2", "C1", or "C2"}

\item{verbose}{A logical value indicating whether to print extra
information. Defaults to FALSE}
}
\value{
A list with two components:
\describe{
 \item{Z1}{A \code{L} x \code{p} matrix, the initial value for the
           \code{L} minor copy numbers of the \code{p} latent features}
 \item{Z2}{A \code{L} x \code{p} matrix, the initial value for the
           \code{L} major copy numbers of the \code{p} latent features}
}
}
\description{
Initialization of the c3co model parameters
}
\examples{
dataAnnotTP <- acnr::loadCnRegionData(dataSet="GSE11976", tumorFrac=1)
dataAnnotN <- acnr::loadCnRegionData(dataSet="GSE11976", tumorFrac=0)
len <- 500*10
nbClones <- 3
bkps <- list(c(100, 250)*10, c(150, 400)*10, c(150, 400)*10)
regions <- list(c("(0,3)", "(0,1)", "(1,2)"),
c("(1,1)", "(0,1)", "(1,1)"), c("(0,2)", "(0,1)", "(1,1)"))
datSubClone <- buildSubclones(len, dataAnnotTP, dataAnnotN,
                              nbClones, bkps, regions)
M <- rbind(c(40, 30, 0),
  c(0, 70, 15),
  c(10, 0, 35),
  c(15, 0, 0),
  c(0, 0, 0))
simu <- mixSubclones(subClones=datSubClone, M)
seg <- segmentData(simu)
res <- initializeZ(seg$Y1, seg$Y2, nb.arch=4)
resC <- initializeZ(seg$Y1+seg$Y2, nb.arch=4)

}
