% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Facetswrapper.R
\name{loadFacetsdata}
\alias{loadFacetsdata}
\title{Function to load Facets data and transform them to c3co format}
\usage{
loadFacetsdata(pathFacets)
}
\arguments{
\item{pathFacets}{The path to load Facets data csv.gz format.}
}
\value{
A data frame under PSCBS format
}
\description{
Function to load Facets data and transform them to c3co format
}
