% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Facetswrapper.R
\name{loadFacetsdata}
\alias{loadFacetsdata}
\title{Loads FACETS data and transforms them to c3co format}
\usage{
loadFacetsdata(path, pattern = "[.]csv[.]gz$")
}
\arguments{
\item{path}{The path to a directory containing FACETS data files
(\file{*.csv.gz}) produced by the \pkg{facets} package.}

\item{pattern}{The filename pattern of files to load.}
}
\value{
A data frame under PSCBS format.
}
\description{
Loads FACETS data and transforms them to c3co format
}
