% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSCBSwrapper.R
\name{loadPSCBSdata}
\alias{loadPSCBSdata}
\title{Loads PSCBS data and transforms them into c3co format}
\usage{
loadPSCBSdata(PSCBSdata)
}
\arguments{
\item{PSCBSdata}{A list that contains PSCBS data.}
}
\value{
A data frame under PSCBS format.
}
\description{
Loads PSCBS data and transforms them into c3co format
}
