% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSCBSwrapper.R
\name{loadPSCBSdata}
\alias{loadPSCBSdata}
\title{Function to load PSCBS data and transform them to InCaSCN format}
\usage{
loadPSCBSdata(pathSegPSCBS)
}
\arguments{
\item{pathSegPSCBS}{The path to load PSCBS data.}
}
\value{
A data frame under PSCBS format
}
\description{
Function to load PSCBS data and transform them to InCaSCN format
}

