% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixSubclone.R
\name{mixSubclones}
\alias{mixSubclones}
\title{Create a mixture of subclones}
\usage{
mixSubclones(subClones, W)
}
\arguments{
\item{subClones}{A list of K subclones, as provided by \code{\link[=buildSubclones]{buildSubclones()}}.
Each subclone is a data.frame with the following columns: \describe{

\item{ct}{total copy number in the tumor}
\item{baft}{B allele fraction in the tumor}
\item{genotype}{germline genotype in 0, 0.5, 1}
\item{cn}{total copy number in the matched normal}
\item{bafn}{B allele fraction in the matched normal}}}

\item{W}{A n x K matrix of weights in \eqn{[0,1]}, as generated by
\code{\link[=rSparseWeightMatrix]{rSparseWeightMatrix()}}.}
}
\value{
A list of n samples obtained by mixing the pure subclones according
to the proportions specified in argument W.
}
\description{
Create a mixture of subclones
}
\details{
The mixture is performed at the scale of minor and major copy
numbers (which is the scale at which copy number events occur). Therefore
the data is first mapped to this scale, then mixed, then mapped back to the
original (tcn, dh) scale.
}
\examples{
dataAnnotTP <- acnr::loadCnRegionData(dataSet="GSE11976", tumorFrac=1)
dataAnnotN <- acnr::loadCnRegionData(dataSet="GSE11976", tumorFrac=0)
len <- 500*10  ## Number of loci
K <- 2L        ## Number of subclones
n <- 3L        ## Number of samples
bkps <- list(c(100, 250)*10, c(150, 400)*10)
regions <- list(c("(0,3)", "(0,2)", "(1,2)"), c("(1,1)", "(0,1)", "(1,1)"))
datSubClone <- buildSubclones(len=len, nbClones=K, bkps=bkps, regions=regions,
                              dataAnnotTP=dataAnnotTP, dataAnnotN=dataAnnotN)
W <- rSparseWeightMatrix(nb.samp=n, nb.arch=K)
mixture <- mixSubclones(datSubClone, W=W)

}
