% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixSubclone.R
\name{mixSubclones}
\alias{mixSubclones}
\title{Create a mixture of subclones}
\usage{
mixSubclones(subClones, W)
}
\arguments{
\item{subClones}{The subclones used to create the mixture ideally by
\code{\link[=buildSubclones]{buildSubclones()}}.}

\item{W}{A matrix of weights in percentage \eqn{[0,100]}
(without fraction of normal cells).}
}
\value{
The mixture.
}
\description{
Create a mixture of subclones
}
\examples{
dataAnnotTP <- acnr::loadCnRegionData(dataSet="GSE11976", tumorFrac=1)
dataAnnotN <- acnr::loadCnRegionData(dataSet="GSE11976", tumorFrac=0)
len <- 500*10  ## Number of loci
K <- 2L        ## Number of subclones
n <- 1L        ## Number of samples
bkps <- list(c(100, 250)*10, c(150, 400)*10)
regions <- list(c("(0,3)", "(0,2)", "(1,2)"), c("(1,1)", "(0,1)", "(1,1)"))
datSubClone <- buildSubclones(len=len, nbClones=K, bkps=bkps, regions=regions,
                              dataAnnotTP=dataAnnotTP, dataAnnotN=dataAnnotN)
W <- rSparseWeightMatrix(nb.samp=n, nb.arch=5L)               
mixture <- mixSubclones(datSubClone, W=W)

}
