% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelFitStats.R
\name{modelFitStatistics}
\alias{modelFitStatistics}
\title{Print statistics of a c3co model}
\usage{
modelFitStatistics(Y, Yhat, What, Zhat)
}
\arguments{
\item{Y}{Original matrix of data.}

\item{Yhat}{Estimates.}

\item{What}{Inferred weight matrix.}

\item{Zhat}{Inferred subclones matrix.}
}
\value{
The main statistics of the inferred model.
}
\description{
Print statistics of a c3co model
}
