% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelFitStats.R
\name{modelFitStatistics}
\alias{modelFitStatistics}
\title{Print statistics of a c3co model}
\usage{
modelFitStatistics(Y, Yhat, What, Zhatt)
}
\arguments{
\item{Y}{Original n-by-J matrix of data.}

\item{Yhat}{Estimated n-by-J matrix.}

\item{What}{Inferred n-by-K weight matrix.}

\item{Zhatt}{The transposed version of the inferred n-by-J subclones
matrix Zhat.}
}
\value{
The main statistics of the inferred model.
}
\description{
Print statistics of a c3co model
}
