% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{posFused-class}
\alias{posFused-class}
\title{Class for the object returned by \code{positiveFusedLasso} function}
\description{
Class for the object returned by \code{positiveFusedLasso} function
}
\section{Slots}{

\describe{
\item{\code{S}}{A list containing the (final) minor, major and total copy number
estimates for the latent features, as inferred by
\code{\link{positiveFusedLasso}}.}

\item{\code{S0}}{idem for the initial estimates.}

\item{\code{W}}{A matrix containing weights inferred by \code{positiveFusedLasso}.}

\item{\code{E}}{A list containing estimates of minor copy number and major copy
number inferred by \code{positiveFusedLasso}.}
}}

