% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{posFused-class}
\alias{posFused-class}
\title{Class for the object returned by \code{\link[=positiveFusedLasso]{positiveFusedLasso()}}}
\description{
Class for the object returned by \code{\link[=positiveFusedLasso]{positiveFusedLasso()}}
}
\section{Slots}{

\describe{
\item{\code{Y}}{A list containing the original minor, major and total copy number
signals.}

\item{\code{Zt}}{A list containing the (final) minor, major and total copy number
estimates for the latent features, as inferred by
\code{\link[=positiveFusedLasso]{positiveFusedLasso()}}.}

\item{\code{Z0t}}{idem for the initial estimates.}

\item{\code{W}}{A matrix containing weights inferred by \code{\link[=positiveFusedLasso]{positiveFusedLasso()}}.}

\item{\code{mu}}{A list containing the intercept vectors inferred by
\code{\link[=positiveFusedLasso]{positiveFusedLasso()}}.}

\item{\code{E}}{A list containing estimates of minor copy number and major copy
number inferred by \code{\link[=positiveFusedLasso]{positiveFusedLasso()}}.}

\item{\code{params}}{A numeric vector, the input parameters.}

\item{\code{failure}}{A logical value indicating whether the estimation worked or
failed (because of the non-invertibility of \eqn{W^tW}).}

\item{\code{converged,iterations}}{Did the model fit converge and how many
iterations did it run?}
}}

