% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{pvePlot}
\alias{pvePlot}
\title{Function to plot PVE}
\usage{
pvePlot(resC3co, bestNbLatent = NULL, ylim = c(0, 1))
}
\arguments{
\item{resC3co}{output from c3co.}

\item{bestNbLatent}{best number of latent profiles.}

\item{ylim}{a vector that define min and max of y-axis}
}
\value{
PVE curve
}
\description{
Function to plot PVE
}

