% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{pvePlot}
\alias{pvePlot}
\title{Plot Percentage Variance Explained (PVE) against number of features}
\usage{
pvePlot(res, bestNbLatent = NULL, ylim = c(0, 1))
}
\arguments{
\item{res}{Result from \code{\link[=c3co]{c3co()}}.}

\item{bestNbLatent}{Best number of latent profiles.}

\item{ylim}{A vector that define min and max of y-axis.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
Plot Percentage Variance Explained (PVE) against number of features
}
