% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{pvePlot2}
\alias{pvePlot2}
\title{Function to plot PVE (low-level)}
\usage{
pvePlot2(dat, bestNbLatent = NULL, ylim = c(0, 1))
}
\arguments{
\item{dat}{A data frame, typically the element \code{best} of the
\code{config} slot from \code{c3co} or \code{fitC3co}}

\item{bestNbLatent}{best number of latent profiles.}

\item{ylim}{a vector that define min and max of y-axis}
}
\value{
PVE curve
}
\description{
Function to plot PVE (low-level)
}
