% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentData.R
\name{segmentData}
\alias{segmentData}
\title{Segmentation function}
\usage{
segmentData(dat, stat = "C1C2")
}
\arguments{
\item{dat}{A list of data frame for each patient. Data frame containing 
\describe{
 \item{tcn}{Total copy number}
 \item{dh}{Mirrored B allele fraction}
 \item{pos}{Position on the genome}
 \item{chr}{Chromosome}
 }}

\item{stat}{"TCN or "C1C2" paramater to segment the data. If \code{stat==TCN}, the segmentation will be done on TCN only.}
}
\value{
Binned Minor and Major copy number with list of breakpoints
}
\description{
Segmentation function
}
\examples{
dataAnnotTP <- acnr::loadCnRegionData(dataSet="GSE11976", tumorFrac=1)
dataAnnotN <- acnr::loadCnRegionData(dataSet="GSE11976", tumorFrac=0)
len <- 500*10
nbClones <- 3
bkps <- list(c(100,250)*10, c(150,400)*10,c(150,400)*10)
regions <-list(c("(0,3)", "(0,2)","(1,2)"), 
c("(1,1)", "(0,1)","(1,1)"), c("(0,2)", "(0,1)","(1,1)"))
datSubClone <- buildSubclones(len, dataAnnotTP, dataAnnotN, nbClones, bkps, regions)
M <- getWeightMatrix(100,0, 3, 15, sparse.coeff=0.7, contam.coeff=0.6, contam.max=2)
dat <- mixSubclones(subClones=datSubClone, M)
res <- segmentData(dat)
res2 <- segmentData(dat, stat="TCN")
}
