% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentData.R
\name{segmentData}
\alias{segmentData}
\title{Joint segmentation}
\usage{
segmentData(dat, stat = c("C1C2", "TCN"), verbose = FALSE)
}
\arguments{
\item{dat}{A list of data frames containing columns:
\describe{
\item{\code{tcn}}{Total copy number}
\item{\code{dh}}{Mirrored B allele fraction}
\item{\code{pos}}{Position on the genome}
\item{\code{chr}}{Chromosome}
}}

\item{stat}{\code{"TCN"} or \code{"C1C2"} parameter to segment the data.
If \code{stat ==}"TCN"`, the segmentation will be done on TCN only.}

\item{verbose}{A logical value indicating whether to print extra
information. Defaults to \code{FALSE}.}
}
\value{
Binned minor and major copy numbers with list of breakpoints.
}
\description{
Joint segmentation by Recursive Binary Segmentation followed by Dynamic
Programming
}
\details{
This function is a wrapper around the \code{\link[jointseg:jointSeg]{jointseg::jointSeg()}} of
the \pkg{jointseg} package.
}
\examples{
dataAnnotTP <- acnr::loadCnRegionData(dataSet="GSE11976", tumorFrac=1)
dataAnnotN <- acnr::loadCnRegionData(dataSet="GSE11976", tumorFrac=0)
len <- 500*10
nbClones <- 3
bkps <- list(c(100, 250)*10, c(150, 400)*10, c(150, 400)*10)
regions <- list(c("(0,3)", "(0,2)", "(1,2)"),
c("(1,1)", "(0,1)", "(1,1)"), c("(0,2)", "(0,1)", "(1,1)"))
datSubClone <- buildSubclones(len, nbClones, bkps, regions, dataAnnotTP, dataAnnotN)
M <- rSparseWeightMatrix(15, 3, sparse.coeff=0.7)
dat <- mixSubclones(subClones=datSubClone, M)
res <- segmentData(dat)
res2 <- segmentData(dat, stat="TCN")

}
\references{
Gey, S., & Lebarbier, E. (2008). Using CART to Detect Multiple
Change Points in the Mean for Large Sample.
http://hal.archives-ouvertes.fr/hal-00327146/

Pierre-Jean, M., Rigaill, G. & Neuvial, P. (2015). Performance
evaluation of DNA copy number segmentation methods.  Briefings in
Bioinformatics 16 (4): 600-615
}
