% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\docType{methods}
\name{showPosFused}
\alias{showPosFused}
\alias{showPosFused,posFused-method}
\title{The method showPosFused}
\usage{
showPosFused(this)

\S4method{showPosFused}{posFused}(this)
}
\arguments{
\item{this}{an object from the following class: [\code{\linkS4class{posFused}}]}
}
\value{
nothing
}
\description{
Print the message that is contained in slots S, W and E.
\itemize{
\item{showPosFused(this)}{this is an object from class  [\code{\linkS4class{posFused}}].}
}
}

