% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{showPosFused}
\alias{showPosFused}
\alias{showPosFused,posFused-method}
\title{Display the content of an object of class posFused}
\usage{
showPosFused(this)

\S4method{showPosFused}{posFused}(this)
}
\arguments{
\item{this}{an object of class: [\code{\linkS4class{posFused}}]}
}
\description{
Display the content of an object of class posFused
}
