% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{Wplot}
\alias{Wplot}
\alias{Wplot,c3coFit-method}
\title{Plot the weight matrix}
\usage{
Wplot(this, idxBest, rownamesW = NULL, col = NULL, margins = c(5, 7),
  posLegend = NA, listPheno, colsPheno, colLegend, labelLegend,
  cexCol = 1.5, ...)

\S4method{Wplot}{c3coFit}(this, idxBest, rownamesW = NULL, col = NULL,
  margins = c(5, 7), posLegend = NA, listPheno, colsPheno, colLegend,
  labelLegend, cexCol = 1.5, ...)
}
\arguments{
\item{this}{An object of class \code{\linkS4class{c3coFit}}}

\item{idxBest}{A integer, the number of latent features}

\item{rownamesW}{A vector that contains identification of patients}

\item{col}{A vector that contains colors for the heatmap}

\item{margins}{A vector margins}

\item{posLegend}{position of the legend to be passed to \code{plot}}

\item{listPheno}{A matrix that contains details on phenotype for each
patient. Could be location or time point of tumors for example}

\item{colsPheno}{Matrix that containts colors for each type of variable
in phenotype}

\item{colLegend}{colors for clinical data}

\item{labelLegend}{labels for clinical data}

\item{cexCol}{size of labels of columns by (default 1.5)}

\item{...}{other paramaters to personalize heatmap (see \code{heatmap.3})}
}
\value{
a heatmap of W
}
\description{
Plot the weight matrix
}
