% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{Wplot}
\alias{Wplot}
\title{Function to plot W}
\usage{
Wplot(dataBest, rownamesW = NULL, col = NULL, margins = c(5, 7),
  posLegend = NA, listPheno, colsPheno, colLegend, labelLegend,
  cexCol = 1.5, ...)
}
\arguments{
\item{dataBest}{A list from output of InCaSCN}

\item{rownamesW}{A vector that contains identification of patients}

\item{col}{A vector that contains colors for the heatmap}

\item{margins}{A vector margins}

\item{posLegend}{position of the legend as for \code{plot}}

\item{listPheno}{A Matrix that contains details on phenotype for each patient. Could be location or time point of tumors for example}

\item{colsPheno}{Matrix that containts colors for each type of variable in phenotype}

\item{colLegend}{colors for clinical data}

\item{labelLegend}{labels for clinical data}

\item{cexCol}{size of labels of columns by (default 1.5)}

\item{...}{other paramaters to personalize heatmap (see \code{heatmap.3.R})}
}
\value{
Heatmap of W
}
\description{
Function to plot W
}

