% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{Zplot}
\alias{Zplot}
\title{Plot latent profiles along chromosomes}
\usage{
Zplot(df, scalePosToMb = FALSE)
}
\arguments{
\item{df}{data.frame object output from \code{createZdf}}

\item{scalePosToMb}{a logical value, should 'x' positions be scaled to
MegaBases? Defaults to FALSE}
}
\value{
plot latent profiles along chromosomes
}
\description{
Plot latent profiles along chromosomes
}
