% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{Zplot}
\alias{Zplot}
\title{Plot latent profiles along chromosomes}
\usage{
Zplot(df, scalePosToMb = FALSE)
}
\arguments{
\item{df}{data.frame Object output from \code{\link[=createZdf]{createZdf()}}.}

\item{scalePosToMb}{A logical, should 'x' positions be scaled to
megabases? Defaults to \code{FALSE}.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
Plot latent profiles along chromosomes
}
