% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{Zplot}
\alias{Zplot}
\title{Plot latent profiles along chromosomes}
\usage{
Zplot(df)
}
\arguments{
\item{df}{data.frame object output from \code{createZdf}}
}
\value{
plot latent profiles along chromosomes
}
\description{
Plot latent profiles along chromosomes
}
