% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{Zplot}
\alias{Zplot}
\title{Function to plot Latent profiles}
\usage{
Zplot(df)
}
\arguments{
\item{df}{data.frame object output from \code{createZdf}}
}
\value{
plot of Latent profiles
}
\description{
Function to plot Latent profiles
}
